RandomRecipe en Detail klasse samengevoegd
