+++ app/src/main/java/com/example/s1103338_s1104467_iiatimd_app/MainActivity.java
@@ -56,36 +56,6 @@ public class MainActivity extends AppCompatActivity {
 //            }
 //        });
 
-        //maakt volleySingleton aan & vraagt requestqueue aan
-//        RequestQueue queue = VolleySingleton.getInstance(this.getApplicationContext()).getRequestQueue();
-//        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET, "https://swapi.dev/api/people/1/", null, new Response.Listener<JSONObject>() {
-//            @Override
-//            public void onResponse(JSONObject response) {
-//                //hier moet komen wat er gaat gebeuren als url is aangevraagd
-//            }
-//        }, new Response.ErrorListener() {
-//            @Override
-//            public void onErrorResponse(VolleyError error) {
-//                Log.d("gefaald", error.getMessage());
-//            }
-//        }
-//        );
-//        StringRequest stringRequest = new StringRequest(Request.Method.GET, "https://www.google.nl", new Response.Listener<String>() {
-//            @Override
-//            public void onResponse(String response) {
-//                Log.d("gelukt", response.substring(0,50));
-//            }
-//        }, new Response.ErrorListener() {
-//            @Override
-//            public void onErrorResponse(VolleyError error) {
-//                Log.d("gefaald", error.getMessage());
-//            }
-//        }
-//        );
-//        VolleySingleton.getInstance(this).addToRequestQueue(jsonObjectRequest);
-
-//        recyclerViewAdapter = new ; vul hier de recycler db in
-//        recyclerView.setAdapter(recyclerViewAdapter);
     }
