+++ app/src/main/java/com/example/s1103338_s1104467_iiatimd_app/MyFirebaseMessagingService.java
@@ -28,7 +28,7 @@ public class MyFirebaseMessagingService extends FirebaseMessagingService {
     }
     @RequiresApi(api = Build.VERSION_CODES.O)
     private void sendNotification(String messageBody){
-        Intent intent = new Intent(this, RandomRecipies.class);
+        Intent intent = new Intent(this, RecipeDetail.class);
         intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
         intent.putExtra("shake", messageBody);
         PendingIntent pendingIntent = PendingIntent.getActivity(this, 0, intent, PendingIntent.FLAG_ONE_SHOT);
